requirements.txt
Django==2.1.2
django-filter==2.0.0
djangorestframework==3.9.0
factory-boy==2.11.1
Faker==0.9.2
Markdown==3.0.1
Pillow==5.3.0
python-dateutil==2.7.5
pytz==2018.6
six==1.11.0
text-unidecode==1.2
